Refactor: Make comments generic using ContentTypes
Replace product-specific comment system with a generic, reusable comment system using Django's ContentTypes framework.

**Major Changes:**
- Move Comment model from shop to core app
- Use GenericForeignKey to allow comments on any model (Product, BlogPost, etc.)
- Generic API endpoints: /comments/{app_label}/{model_name}/{identifier}/
- Backward-compatible: Products can use /comments/shop/product/{slug}/
- Future-ready: Blog posts can use /comments/blog/blogpost/{slug}/

**New API Structure:**
- GET/POST /comments/{app_label}/{model_name}/{identifier}/ - List/Create comments
- DELETE /comments/{comment_id}/ - Delete a comment

**Examples:**
- Product comments: /comments/shop/product/my-product-slug/
- Blog comments: /comments/blog/blogpost/my-blog-post-slug/

**Features Preserved:**
- Auto-approval for verified users
- Anonymous comment support
- Soft delete
- Pagination (20 items/page)
- Admin interface with bulk actions
- Full moderation capabilities

**Migration:**
- core.migrations.0003_comment - Creates generic Comment table
- Removes shop-specific ProductComment model and views
