+++ shop/models/__init__.py
@@ -11,7 +11,6 @@ from .category import Category
 from .order import Order, OrderItem, ShippingInfo, PaymentInfo
 from .promo import PromoCode
 from .payment import Payment, PaymentTransaction
-from .comment import ProductComment
 
 __all__ = [
     "BaseModel",
@@ -28,5 +27,4 @@ __all__ = [
     "Payment",
     "PaymentTransaction",
     "ProductInfoPage",
-    "ProductComment",
 ]
